+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Added the anchor to the crop insurance county level maps and reuseable function for the county-level maps in the future [#445](https://github.com/policy-design-lab/pdl-frontend/issues/445)
+
 ## [1.15.0] - 2026-03-04
 
 ### Added