+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.16.0] - 2026-03-26
+
+### Added
+- Added the anchor to the crop insurance county level maps and reuseable function for the county-level maps in the future [#445](https://github.com/policy-design-lab/pdl-frontend/issues/445)
+
 ## [1.15.0] - 2026-03-04
 
 ### Added
@@ -424,6 +429,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Map data json [#12](https://github.com/policy-design-lab/pdl-frontend/issues/12)
 - Final landing page changes for initial milestone [#15](https://github.com/policy-design-lab/pdl-frontend/issues/15)
 
+[1.16.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.15.0...1.16.0
 [1.15.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.14.0...1.15.0
 [1.14.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.13.0...1.14.0
 [1.13.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.12.0...1.13.0